+++ src/Service/CronTaskHandler.php
@@ -10,6 +10,7 @@ use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\MessageQueue\ScheduledTask\ScheduledTaskHandler;
 use Shopware\Core\Kernel;
+use Shopware\Core\System\StateMachine\Aggregation\StateMachineTransition\StateMachineTransitionActions;
 use Shopware\Core\System\SystemConfig\SystemConfigService;
 use Symfony\Bridge\Monolog\Logger;
 use Symfony\Contracts\Translation\TranslatorInterface;
@@ -23,6 +24,9 @@ class CronTaskHandler extends ScheduledTaskHandler
     private EntityRepositoryInterface $customerRepository;
     private OrderTransactionStateHandler $transactionStateHandler;
     private TranslatorInterface $translator;
+
+    const CANCELLATION_MODE = 'cancellation';
+    const CAPTURE_MODE = 'captrure';
 
     public function __construct(
         EntityRepositoryInterface    $scheduledTaskRepository,
@@ -55,71 +59,105 @@ class CronTaskHandler extends ScheduledTaskHandler
         $salesChannels = $this->salesChannelRepository->search(new Criteria(), Context::createDefaultContext());
         foreach ($salesChannels as $salesChannel) {
 
-            $ordersList = $this->getOrderList($salesChannel->getId());
-
-            if (empty($ordersList)) {
-                continue;
-            }
-
-            foreach ($ordersList as $order) {
+            $captureOrdersList = $this->getOrderList($salesChannel->getId(), self::CAPTURE_MODE);
+            foreach ($captureOrdersList as $order) {
                 $this->processOrder($order);
+            }
+            $cancellationOrdersList = $this->getOrderList($salesChannel->getId(), self::CANCELLATION_MODE);
+            foreach ($cancellationOrdersList as $order) {
+                $transactionId = strtolower($order['trans_id']);
+                $this->transactionStateHandler->cancel($transactionId, Context::createDefaultContext());
             }
         }
     }
 
     /**
-     * @param $salesChannelId
-     * @return array|\mixed[][]|void
+     * @param string $salesChannelId
+     * @param string $mode
+     * @return array
      * @throws \Doctrine\DBAL\Driver\Exception
      * @throws \Doctrine\DBAL\Exception
      */
-    private function getOrderList($salesChannelId)
+    private function getOrderList(string $salesChannelId, string $mode): array
     {
         $adapter = new WorldlineSDKAdapter($this->systemConfigService, $this->logger, $salesChannelId);
-        $captureConfig = $adapter->getPluginConfig(Form::AUTO_CAPTURE);
-        if ($captureConfig === Form::AUTO_CAPTURE_DISABLED) {
-            return;
-        }
-
-        $daysInterval = $this->getDaysInterval($captureConfig);
         $connection = Kernel::getConnection();
 
         $qb = $connection->createQueryBuilder();
-        $qb->select('ot.custom_fields, ot.updated_at')
+
+        $qb->select('HEX(o.id) as id, o.custom_fields, o.updated_at')
             ->from('`order`', 'o')
-            ->leftJoin('o', 'order_transaction', 'ot', "ot.order_id = o.id")
             ->where("o.sales_channel_id = UNHEX(:salesChannelId)")
-            ->andWhere(
-                $qb->expr()->or(
-                    $qb->expr()->like('ot.custom_fields', "'%payment_transaction_status\": \"5%'"),
-                    $qb->expr()->like('ot.custom_fields', "'%payment_transaction_status\": \"56%'")
-                )
-            )
-            ->orderBy('ot.updated_at', 'DESC')
+            ->orderBy('o.updated_at', 'DESC')
             ->setParameter('salesChannelId', $salesChannelId);
 
-        if ($daysInterval > 0) {
-            $qb->andWhere("DATEDIFF(NOW(), ot.updated_at) > :daysInterval")
-                ->setParameter('daysInterval', $daysInterval);
+        switch ($mode) {
+            case "capture":
+            {
+                $captureConfig = $adapter->getPluginConfig(Form::AUTO_CAPTURE);
+                if ($captureConfig === Form::AUTO_PROCESSING_DISABLED) {
+                    return [];
+                }
+
+                $qb->andWhere(
+                    $qb->expr()->or(
+                        $qb->expr()->like('o.custom_fields', "'%payment_transaction_status\": \"5%'"),
+                        $qb->expr()->like('o.custom_fields', "'%payment_transaction_status\": \"56%'")
+                    )
+                );
+
+                $timeInterval = $this->getTimeInterval($captureConfig) * 60 * 60 * 24;
+                break;
+            }
+            case "cancellation":
+            {
+                $cancellationConfig = $adapter->getPluginConfig(Form::AUTO_CANCEL);
+                if ($cancellationConfig === Form::AUTO_PROCESSING_DISABLED) {
+                    return [];
+                }
+
+                $qb->select('o.custom_fields, hex(ot.id) as trans_id, sms.technical_name')
+                    ->leftJoin('o', 'order_transaction', 'ot', "ot.order_id = o.id")
+                    ->leftJoin('ot', 'state_machine_state', 'sms', "sms.id = ot.state_id")
+                    ->andWhere(
+                        $qb->expr()->or(
+                            $qb->expr()->like('o.custom_fields', "'%payment_transaction_status\": \"0%'"),
+                            $qb->expr()->like('o.custom_fields', "'%payment_transaction_status\": \"46%'")
+                        )
+                    )
+                    ->andWhere("sms.technical_name != :technicalName")
+                    ->setParameter('technicalName', StateMachineTransitionActions::ACTION_CANCEL)
+                ;
+
+                $timeInterval = $this->getTimeInterval($cancellationConfig) * 60 * 60;
+                break;
+            }
+            default:
+                return [];
         }
 
-        return $qb->execute()->fetchAllAssociative();
+        if ($timeInterval > 0) {
+            $qb->andWhere("UNIX_TIMESTAMP(NOW()) - UNIX_TIMESTAMP(o.updated_at) > :timeInterval")
+                ->setParameter('timeInterval', $timeInterval);
+        }
+
+        return $qb->execute()->fetchAllAssociative() ? : [];
     }
 
     /**
      * @return int
      */
-    private function getDaysInterval($captureConfig)
+    private function getTimeInterval($config)
     {
-        return (int)filter_var($captureConfig, FILTER_SANITIZE_NUMBER_INT);
+        return (int)filter_var($config, FILTER_SANITIZE_NUMBER_INT);
     }
 
     /**
-     * @param $order
+     * @param array $order
      * @return void
      * @throws \Exception
      */
-    private function processOrder($order)
+    private function processOrder(array $order)
     {
         if (!array_key_exists('custom_fields', $order)) {
             return;